requirements.txt
telethon
python-dotenv
psycopg2-binary
pandas
dbt-postgres